requirements.txt
srcomapi
discord.py